✅ Add test for file list filter bar on different window width
